+++ tests/mutable.lua
@@ -1,3 +1,0 @@
-TestMutable = {
-	["test: TODO"] = function () error "Not implemented" end;
-}