+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # CHANGELOG
 ## Unreleased
+- Add mutable array
 
 ## [1.0.0](../../tree/1.0.0) - XXXX-XX-XX
 Release